requirement.txt
APScheduler==3.6.3
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.41.0
django-environ==0.4.5
django-rest-auth==0.9.5
django-widget-tweaks==1.4.1
djangorestframework==3.9.1
html5lib==1.0.1
idna==2.9
oauthlib==3.1.0
olefile==0.44
Pillow==5.4.1
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2017.2
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tzlocal==2.1
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
